backend/requirements.txt
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
black==25.1.0
Brotli==1.1.0
cffi==1.17.1
click==8.2.1
colorlog==6.9.0
cssselect2==0.8.0
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
fonttools==4.59.2
h11==0.16.0
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dotenv==1.1.1
redis==6.4.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.14.1
uvicorn==0.35.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
